+++ src/main/java/com/codestates/pre032/pre032/question/QuestionController.java
@@ -2,7 +2,11 @@ package com.codestates.pre032.pre032.question;
 
 import com.codestates.pre032.pre032.answer.AnswerController;
 import com.codestates.pre032.pre032.dto.MainResponseDto;
+import com.codestates.pre032.pre032.exception.UnauthorizedException;
+import com.codestates.pre032.pre032.security.dto.TokenDto;
 import com.codestates.pre032.pre032.tag.TagService;
+import com.codestates.pre032.pre032.user.User;
+import com.codestates.pre032.pre032.user.UserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -21,12 +25,15 @@ public class QuestionController {
 
     private final AnswerController answerController;
 
+    private final UserService userService;
+
     private final TagService tagService;
 
-    public QuestionController(QuestionService questionService, QuestionMapper questionMapper, AnswerController answerController, TagService tagService) {
+    public QuestionController(QuestionService questionService, QuestionMapper questionMapper, AnswerController answerController, UserService userService, TagService tagService) {
         this.questionService = questionService;
         this.questionMapper = questionMapper;
         this.answerController = answerController;
+        this.userService = userService;
         this.tagService = tagService;
     }
 
@@ -34,13 +41,31 @@ public class QuestionController {
     @PostMapping("/add")
 //    @PreAuthorize("isAuthenticated()")
     public ResponseEntity addQuestion(@Validated @RequestBody QuestionDto.Post requestBody) {
+        if (requestBody.getAccessToken()==null){
+            throw new UnauthorizedException("로그인이 필요합니다.");
+        }
+
+        User user = userService.findByAccessToken(requestBody.getAccessToken());
+
         Question question = questionService.create(questionMapper.questionPostDtoToQuestion(requestBody),
-                requestBody.getTags());
+                requestBody.getTags(),user);
         QuestionDto.Response response = questionMapper.questionToQuestionResponse(question);
 
         return new ResponseEntity<>(response, HttpStatus.CREATED);
     }
-
+    // 수정 페이지 요청
+    @GetMapping("/{questionId}/edit")
+    public ResponseEntity getUpdatePage(@PathVariable("questionId") Long id,@RequestBody TokenDto requestBody){
+        //로그인이 안되어있으면
+        if (requestBody.getAccessToken()==null){
+            throw new UnauthorizedException("로그인이 필요합니다.");
+        }
+        // 본인이 작성한 질문이 아니면
+        if (!questionService.hasQuestion(id,userService.findByAccessToken(requestBody.getAccessToken()))){
+            throw new UnauthorizedException("작성자가 아닙니다.");
+        }
+        return new ResponseEntity(HttpStatus.OK);
+    }
     // 수정 기능
     @PatchMapping("/{questionId}/edit")
     public ResponseEntity updateQuestion(@PathVariable("questionId") Long id,
@@ -70,7 +95,12 @@ public class QuestionController {
 
     // 질문 삭제 기능
     @DeleteMapping("/{questionId}/delete")
-    public ResponseEntity deleteQuestion(@PathVariable("questionId") Long id) {
+    public ResponseEntity deleteQuestion(@PathVariable("questionId") Long id,
+                                         @RequestBody TokenDto requestBody) {
+        // 본인이 작성한 질문이 아니면
+        if (!questionService.hasQuestion(id,userService.findByAccessToken(requestBody.getAccessToken()))){
+            throw new UnauthorizedException("작성자가 아닙니다.");
+        }
         Question question = this.questionService.find(id);
         this.questionService.delete(id);
         return new ResponseEntity(HttpStatus.OK);
@@ -98,21 +128,5 @@ public class QuestionController {
         List<Question> questions = questionService.sortCount();
         return new ResponseEntity<>(
                 new MainResponseDto(questionMapper.questionsToQuestionContentResponsesDto(questions)), HttpStatus.OK);
-    }
-
-//    @GetMapping("/sortByViewCount")
-//    public ResponseEntity findTitle(){
-//        List<Question> questions = questionService.findTitle("스프링 에러입니다");
-//        return new ResponseEntity<>(
-//                new MainResponseDto(questionMapper.questionsToQuestionContentResponsesDto(questions)), HttpStatus.OK);
-//    }
-
-
-
-    // 테스트용 메서드 모든 질문 삭제
-    @PostMapping("/deleteAll")
-    public ResponseEntity deleteAll() {
-        questionService.deleteAll();
-        return new ResponseEntity<>(HttpStatus.OK);
     }
 }
